% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.downsample.R
\name{aoi.downsample-deprecated}
\alias{aoi.downsample-deprecated}
\title{aoi.downsample}
\usage{
aoi.downsample(x, bin.length = NULL, aoi_col = c())
}
\arguments{
\item{x}{dataframe}

\item{bin.length}{Length of bins to average}

\item{aoi_col}{Column names that contain aoi measures}
}
\description{
Downsample aoi data
}
\details{
This function will reduce the sampling fequency
}
\seealso{
\code{\link{pupillometry-deprecated}}
}
\keyword{internal}
