% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.downsample.R
\name{aoi.downsample}
\alias{aoi.downsample}
\title{Downsample aoi data}
\usage{
aoi.downsample(x, bin.length = NULL, aoi_col = c())
}
\arguments{
\item{x}{dataframe}

\item{bin.length}{Length of bins to average}

\item{aoi_col}{Column names that contain aoi measures}
}
\description{
This function will reduce the sampling fequency
}
\examples{
aoi.downsample(x, bin.length = 100)
}
\keyword{downsample}
