% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.server.R
\name{app.server}
\alias{app.server}
\title{app.server}
\usage{
app.server()
}
\value{
A function.
}
\description{
Server function for the pupillometry app. When using the app locally, use `pupillometry.app()` instead.
}
