% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinecorrect.R
\name{baselinecorrect}
\alias{baselinecorrect}
\title{Pupil Data Function}
\usage{
baselinecorrect(x, bc.duration = 200, start.target = "",
  iteration = 1)
}
\arguments{
\item{x}{dataframe}

\item{bc.duration}{Duration pre-trial baseline to use for correction}

\item{start.target}{Stimulus that identifies the target onset}

\item{iteration}{the nth time this data has been baseline corrected}
}
\description{
This function applies a pre-trial baseline correction on the data
}
\examples{
baselinecorrect(file = "path/filename", baseline.duration = 2000, start.trial = "# Message: Target")
}
\keyword{baseline}
