% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Pupil Preprocessing Function}
\usage{
downsample(x, bin.length = "")
}
\arguments{
\item{x}{dataframe}

\item{bin.length}{Length of bins to average}
}
\description{
This function will reduce the sampling fequency
}
\examples{
downsample(x, bin.length = 100)
}
\keyword{downsample}
