% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye.method.R
\name{eye.method}
\alias{eye.method}
\title{Preprocessnig Function}
\usage{
eye.method(x, eye.recorded = "", method = "average", cor.criteria = 0.9)
}
\arguments{
\item{x}{dataframe}

\item{eye.recorded}{Is there pupil data for the left, right, or both eyes?}

\item{method}{Which method of using left and right eye data? Average, missing, left, or right.}

\item{cor.criteria}{If using the average method, then what is the criteria of how highly the two eyes need to correlate?}
}
\description{
This function is for preprocessing baseline pupil data. It takes as input a data file that has been processed
with the .convert() functions. The output of this function is a preprocessed data file. Preprocessing steps can
include interpolation and smoothing
}
\examples{
eye.method(x)
}
\keyword{eye}
