% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getQualitativePalette}
\alias{getQualitativePalette}
\title{getQualitativePalette}
\usage{
getQualitativePalette(nbcolors)
}
\arguments{
\item{nbcolors}{number of colors (from 1 to 22)}
}
\value{
A vector of colors
}
\description{
Returns a qualitative color palette of the given size. If less than 23 colors are required,
the colors are based on Paul Tol's palettes. If more, the `randomcoloR` package is used.
}
