% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getTimeDriftCoefficient}
\alias{getTimeDriftCoefficient}
\title{getTimeDriftCoefficient}
\usage{
getTimeDriftCoefficient(dat, responseBins)
}
\arguments{
\item{dat}{A list of data.frames containing at least the `Time` and `Diameter`}

\item{responseBins}{A matrix of response bins (each row with start and end of the bin)}
}
\value{
A numeric vector of the same length as `dat`
}
\description{
Return the coefficient of a Diameter~Time rlm on non-response timepoints
}
