% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latency.R
\name{latency}
\alias{latency}
\title{Pupil DV Calculation}
\usage{
latency(x)
}
\arguments{
\item{x}{vector}
}
\description{
This function returns the latency of max pupil dilation
}
\examples{
latency(x)
}
\keyword{latency}
