% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxtomin.R
\name{maxtomin}
\alias{maxtomin}
\title{Pupil DV Calculation}
\usage{
maxtomin(x)
}
\arguments{
\item{x}{vector}
}
\description{
This function returns the difference between max and min values
}
\examples{
maxtomin(x)
}
\keyword{max}
