% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.consecutive.R
\name{missing.consecutive}
\alias{missing.consecutive}
\title{Missing Data Function}
\usage{
missing.consecutive(x)
}
\arguments{
\item{x}{vector}
}
\description{
This function returns the longest consecutive string of missing data in an array as a proportion of the total number of samples in the array
}
\examples{
missing.consecutive(x)
}
\keyword{missing}
