% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.total.R
\name{missing.total}
\alias{missing.total}
\title{Missing Data Function}
\usage{
missing.total(x)
}
\arguments{
\item{x}{vector}
}
\description{
This function returns the total number of missing data in an array as a proportion of the total number of samples in the array
}
\examples{
missing.total(x)
}
\keyword{missing}
