% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normDiameters}
\alias{normDiameters}
\title{normDiameters}
\usage{
normDiameters(dat, bins, tf = 1, normalizeToBaseline = TRUE,
  normDrift = "no")
}
\arguments{
\item{dat}{A list of datasets, each being a data.frame with at least the `Time` and `Diameter` columns.}

\item{bins}{A list with slots `baseline` and `response`, each corresponding a matrix of bins (one per row, start-end)}

\item{tf}{A time scaling factor (default 1, i.e. no scaling)}

\item{normalizeToBaseline}{Logical; whether to normalize to the baseline (Default TRUE)}

\item{normDrift}{Normalize baseline drift across time. Use 'no' to disable (default), 'global' to identify the drift
across all samples, or 'specific' to correct for sample-specific drift.}
}
\value{
A list of normalized datasets.
}
\description{
normDiameters
}
