% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.count.R
\name{peaks.count}
\alias{peaks.count}
\title{Pupil DV Calculation}
\usage{
peaks.count(x)
}
\arguments{
\item{x}{vector}
}
\description{
This function returns the number of max pupil dilation values
}
\examples{
peaks.count(x)
}
\keyword{peaks}
