% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.distance.R
\name{peaks.dist}
\alias{peaks.dist}
\title{Pupil DV Calculation}
\usage{
peaks.dist(x)
}
\arguments{
\item{x}{vector}
}
\description{
This function returns the distance between max pupil dilation values
}
\examples{
peaks.dist(x)
}
\keyword{peaks}
