% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.comparison.R
\name{plot.comparison}
\alias{plot.comparison}
\title{Testing Function}
\usage{
plot.comparison(import, trial)
}
\arguments{
\item{import}{Import directory}

\item{trial}{Which trial to filter on}
}
\description{
This function will print ggplot figures of na.removed, method.first = interpolate, and method.first = smooth
}
