% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Visually compare timeseries that used different preprocessing mehtods/parameters}
\usage{
plot_comparison(import, files = c(), labels = c(), trial, title = "",
  legend.title = "Labels")
}
\arguments{
\item{import}{Directory with preprocessed data files for comparison}

\item{files}{Vector of files to be compared}

\item{labels}{Vector of labels for plotting.}

\item{trial}{Which trial to filter on}

\item{title}{Title the graph}

\item{legend.title}{Title the legend}
}
\description{
This function will print ggplot figures of na.removed, method.first = interpolate, and method.first = smooth
}
\keyword{plot}
