% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pupil Data Function}
\usage{
preprocess(import = "", pattern = "*.txt", export = "", taskname = "",
  eyetracker = "", trial.start = "", eye.recorded = "",
  eyelink.include = "Time", downsample.Hz = "", ms.conversion = "",
  trialonset.message = "", interpolate = FALSE, smooth = FALSE,
  smooth.window = 5, eye.use = "average", bc = FALSE, bc.duration = 200,
  pretrial.duration = 0, targetonset.message = "", subj.prefix = "-")
}
\arguments{
\item{import}{Folder path to raw data files}

\item{pattern}{Pattern to look for in data files}

\item{export}{Folder path to export preprocessed data to}

\item{eyetracker}{Which eye-tracker was used to record data}

\item{trial.start}{Message used in SMI experiment to mark StartTracking inline}

\item{eye.recorded}{Do you want to inclue the "left", "right', or "both" eyes?}

\item{eyelink.include}{Which columns in the raw data export file do you want to keep}

\item{downsample.Hz}{The frequency you want to downsample to}

\item{ms.conversion}{Conversion factor to convert timing to milliseconds}

\item{trialonset.message}{Message string that marks the start of a trial}

\item{interpolate}{Do you want to do a linear interpolation over missing values?}

\item{smooth}{Do you want to apply a moving average smoothing function (window = 5)?}

\item{smooth.window}{Window size of smoothing function default is 5}

\item{eye.use}{Which method of using left and right eye data? Average, missing, left, or right.}

\item{bc.duration}{PreTarget duration to use for baseline correction}

\item{pretrial.duration}{Duration of pre-trial baseline period in milliseconds}

\item{targetonset.message}{Message string that marks the target onset for baseline correction}

\item{subj.prefix}{The prefix that comes before the subject number in the data file (including "-")}
}
\description{
This function will perform preprocessing on an entire folder of data files. It takes as input a specified folder path
and outputs the preprocessed data to a specified folder path
}
\examples{


}
\keyword{preprocess}
