% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.baselinecorrect.R
\name{pupil.baselinecorrect-deprecated}
\alias{pupil.baselinecorrect-deprecated}
\title{Apply baseline correction}
\usage{
pupil.baselinecorrect(x, baselineoffset.message = "", bc.duration = 200,
  bc.type = "subtractive", match = "exact")
}
\arguments{
\item{x}{dataframe}

\item{baselineoffset.message}{Message string(s) that marks the offset of baseline period(s)}

\item{bc.duration}{Duration baseline period(s) to use for correction}

\item{bc.type}{Do you want to use "subtractive" or "divisive" baseline correction? (default: "subtractive")}

\item{match}{Should the message string be an "exact" match or a "pattern" match?}
}
\description{
This function applies a pre-trial baseline correction on the data
}
\seealso{
\code{\link{pupillometry-deprecated}}
}
\keyword{internal}
