% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.cor.R
\name{pupil.cor-deprecated}
\alias{pupil.cor-deprecated}
\title{Add column that is the correlation between left and rigth eyes}
\usage{
pupil.cor(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function is for preprocessing baseline pupil data. It takes as input a data file that has been processed
with the .convert() functions. The output of this function is a preprocessed data file. Preprocessing steps can
include interpolation and smoothing
}
\seealso{
\code{\link{pupillometry-deprecated}}
}
\keyword{internal}
