% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.eye.R
\name{pupil.eye}
\alias{pupil.eye}
\title{Preprocessnig Function}
\usage{
pupil.eye(x, eye.recorded = "", eye.use = "", missing.allowed = 1)
}
\arguments{
\item{x}{xframe}

\item{eye.recorded}{Is there pupil x for the left, right, or both eyes?}

\item{eye.use}{Which eye to use? Left or right}

\item{missing.allowed}{What proportion of missing data is allowed, per trial?}
}
\description{
If both eyes were recorded from, then this function will correlate the timeseries pupil data from both eyes
and select only one eye data to keep for further preprocessing and output.
This function will also remove trials with too much missing data, specified by `missing.allowed`.
}
\examples{

}
\keyword{pupil}
