% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.interpolate.R
\name{pupil.interpolate}
\alias{pupil.interpolate}
\title{Pupil Preprocessing Function}
\usage{
pupil.interpolate(x, eye.recorded = "")
}
\arguments{
\item{x}{dataframe}

\item{eye.recorded}{Is there pupil data for the left, right, or both eyes?}
}
\description{
This function performs a linear interpolation over missing values.
The 'zoo' package is a dependency
}
\examples{
pupil.interpolate(x)
}
\keyword{interpolate}
