% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.missing.R
\name{pupil.missing}
\alias{pupil.missing}
\title{Preprocessnig Function}
\usage{
pupil.missing(x, eye.recorded = "")
}
\arguments{
\item{x}{dataframe}

\item{eye.recorded}{Is there pupil data for the left, right, or both eyes?}
}
\description{
This function is for preprocessing baseline pupil data. It takes as input a data file that has been processed
with the .convert() functions. The output of this function is a preprocessed data file. Preprocessing steps can
include interpolation and smoothing
}
\examples{
pupil.missing(x)
}
\keyword{preprocess}
