% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.smooth.R
\name{pupil.smooth-deprecated}
\alias{pupil.smooth-deprecated}
\title{Apply smoothing/low frequency filter}
\usage{
pupil.smooth(x, type = "hann", window = 5, hz = "")
}
\arguments{
\item{x}{xframe}

\item{type}{The type of smoothing function to apply. Hann or moving window average}

\item{window}{The size of the smoothing window (default = 5). Value is in milliseconds}

\item{hz}{The recording frequency (used to calculate window size)}
}
\description{
This function performs a smoothing function over an array of x
The 'zoo' package is a dependency
}
\seealso{
\code{\link{pupillometry-deprecated}}
}
\keyword{internal}
