% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.smooth.R
\name{pupil.smooth}
\alias{pupil.smooth}
\title{Pupil Preprocessing Function}
\usage{
pupil.smooth(x, eye.recorded = "", window = 5)
}
\arguments{
\item{x}{dataframe}

\item{eye.recorded}{Is there pupil data for the left, right, or both eyes?}

\item{window}{The size of the smoothing window (default = 5)}
}
\description{
This function performs a smoothing function over an array of data
The 'zoo' package is a dependency
}
\examples{
pupil.smooth(x)
}
\keyword{smooth}
