% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_cor.R
\name{pupil_cor}
\alias{pupil_cor}
\title{Correlate pupil values}
\usage{
pupil_cor(x)
}
\arguments{
\item{x}{dataframe.}
}
\description{
Calculate the correlation between left and right pupil values. See
https://dr-jt.github.io/pupillometry/index.html for more information.
}
\section{Output}{


Adds a column `Pupil.r`
}

