% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_deblink.R
\name{pupil_deblink}
\alias{pupil_deblink}
\title{Deblink}
\usage{
pupil_deblink(
  x,
  extend = 0,
  min_blink_duration = 0,
  plot = FALSE,
  plot_trial = "all"
)
}
\arguments{
\item{x}{dataframe.}

\item{extend}{How many milliseconds to extend blinks
before and after blink detection.}

\item{min_blink_duration}{Minimum duration of a blink in milliseconds.}

\item{plot}{Logical. Inspect a plot of how pupil values changed?}

\item{plot_trial}{what trial(s) to plot. default = "all"}
}
\description{
Deblink pupil data with the option to extend blinks before and after
blink detection. See
https://dr-jt.github.io/pupillometry/ for more information.
}
\section{Output}{


Changes values in column containing pupil data.
}

\section{Plot inspection}{


To inspect how the preprocesing step changed pupil size values,
use `plot = TRUE`.

Warning: this will create a separate plot for every trial and therefore can
be time consuming and overwhelming. The plot argument is meant for initial
exploratory steps to determine the appropriate preprocessing parameters.
}

