% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_downsample.R
\name{pupil_downsample}
\alias{pupil_downsample}
\title{Down-sample pupil data}
\usage{
pupil_downsample(x, bin_length = NULL, id = "Subject", bin.length = NULL)
}
\arguments{
\item{x}{dataframe}

\item{bin_length}{Length of bins to average}

\item{id}{Column name that identifies subject number}

\item{bin.length}{deprecated. see bin_length}
}
\description{
This function will reduce the sampling fequency
}
