% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_inspect.R
\name{pupil_inspect}
\alias{pupil_inspect}
\title{Plot and inspect how preprocessing steps are changing the pupil data.
An exploratory function only.}
\usage{
pupil_inspect(x, aggregate = NULL)
}
\arguments{
\item{x}{dataframe}

\item{aggregate}{column name to aggregate the plots by.
For individual trial plots leave as NULL.}
}
\description{
This function will print ggplot figures of either trial level or aggregated pupil data
}
