% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_merge.R
\name{pupil_merge}
\alias{pupil_merge}
\title{Merge multiple files together}
\usage{
pupil_merge(path = "", pattern = "", output_file = "")
}
\arguments{
\item{path}{Folder location of files to be merged}

\item{pattern}{Pattern to identify files to be merged}

\item{output_file}{File name and path to be saved to.}
}
\description{
This function merges multiple files together.
}
