% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_missing.R
\name{pupil_missing}
\alias{pupil_missing}
\title{Evaluate the amount of missing data}
\usage{
pupil_missing(x, missing.allowed = 1, velocity = "", margin = "")
}
\arguments{
\item{x}{dataframe}

\item{missing.allowed}{What proportion of missing data is allowed, per trial?}

\item{velocity}{The velocity threshold for Blink detection}

\item{margin}{The margin before and after Blink onset and offset}
}
\description{
This function is for preprocessing baseline pupil x. It takes as input a x file that has been processed
with the .convert() functions. The output of this function is a preprocessed x file. Preprocessing steps can
include interpolation and smoothing
}
\examples{
pupil_missing(x)
}
\keyword{preprocess}
