% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_preprocess.R
\name{pupil_preprocess}
\alias{pupil_preprocess}
\title{Preprocess pupil data using this single function}
\usage{
pupil_preprocess(import.dir = NULL, pattern = "*.txt",
  taskname = NULL, subj.prefix = NULL, subj.suffix = NULL,
  timing.file = NULL, output.dir = NULL, output.steps = FALSE,
  eyetracker = NULL, hz = NULL, eye.use = NULL,
  px_to_mm.conversion = NULL, start_tracking.message = "default",
  start_tracking.match = "exact", trial_onset.message = NULL,
  trial_onset.match = "exact", deblink.extend = 0,
  pre_trial.duration = NULL, missing.allowed = 1, interpolate = NULL,
  interpolate.maxgap = Inf, smooth = NULL, smooth.window = 5,
  method.first = NULL, bc = NULL, pre_bc.duration = NULL,
  bc_onset.message = NULL, bc_onset.match = "exact",
  subset = "default", trial.exclude = c(), files.merge = FALSE,
  starttracking.message = NULL, starttracking.match = NULL,
  trialonset.message = NULL, trialonset.match = NULL,
  pretrial.duration = NULL, prebc.duration = NULL,
  bconset.message = NULL, bconset.match = NULL)
}
\arguments{
\item{import.dir}{Folder path to raw data files}

\item{pattern}{Pattern to look for in data files}

\item{taskname}{Name of task - to be used in naming pre-processed files}

\item{subj.prefix}{The unique pattern prefix (letter(s) and/or symbol(s))
that comes before the subject number in the data file}

\item{subj.suffix}{The unique pattern suffix (letter(s) or symbol(s))
that comes after the subject number in the data file}

\item{timing.file}{File location and name that contains timing
information for message markers}

\item{output.dir}{Folder path to output preprocessed data to}

\item{output.steps}{Output files for each step in preprocessing?}

\item{eyetracker}{Which eye-tracker was used to record data}

\item{hz}{At which frequency was pupil data sampled at?
(only required for interpolation and smoothing)}

\item{eye.use}{Which eye to use? Left or right}

\item{px_to_mm.conversion}{The conversion factor to go from
px pupil diameter to mm pupil diameter}

\item{start_tracking.message}{Message used in SMI experiment
to mark StartTracking inline}

\item{start_tracking.match}{Should the message string be an "exact"
match or a "pattern" match?}

\item{trial_onset.message}{Message string that marks the start of a trial}

\item{trial_onset.match}{Should the message string be an "exact"
match or a "pattern" match}

\item{deblink.extend}{How many milliseconds to extend blinks
before and after blink detection}

\item{pre_trial.duration}{Duration of pre-trial baseline period
in milliseconds}

\item{missing.allowed}{What proportion of missing data is allowed,
on a trial-by-trial basis? (Default: 1)}

\item{interpolate}{What type of interpolation to use? linear or cubic-spline}

\item{interpolate.maxgap}{Maximum number of NAs to interpolate over.
Anything gaps over this value will not be interpolated.}

\item{smooth}{The type of smoothing function to apply.
hann or moving window average (mwa)}

\item{smooth.window}{Window size of smoothing function
default is 5 milliseconds}

\item{method.first}{Should "smooth" or "interpolate" be applied first?
(default: NULL)}

\item{bc}{Do you want to use "subtractive" or "divisive"
baseline correction? (default: "subtractive")}

\item{pre_bc.duration}{Duration baseline period(s) to use for correction}

\item{bc_onset.message}{Message string(s) that marks the offset of
baseline period(s)}

\item{bc_onset.match}{Message string(s) that marks the offset of
baseline period(s)}

\item{subset}{Which columns in the raw data output file do you want to keep}

\item{trial.exclude}{Specify if ther are any trials to exclude. Trial number}

\item{files.merge}{Do you want to create a single merge output file?}

\item{starttracking.message}{See start_tracking.message}

\item{starttracking.match}{See start_tracking.match}

\item{trialonset.message}{See trial_onset.message}

\item{trialonset.match}{See trial_onset.match}

\item{pretrial.duration}{See pre_trial.duration}

\item{prebc.duration}{See pre_bc.duration}

\item{bconset.message}{See bc_onset.duration}

\item{bconset.match}{See bc_onset.match}
}
\description{
This function will perform preprocessing on an entire folder of data files.
It takes as input a specified folder path
and outputs the preprocessed data to a specified folder path
}
