% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_read.R
\name{pupil_read}
\alias{pupil_read}
\title{Convert "messy" raw data file to "tidy" raw data file}
\usage{
pupil_read(
  file,
  eyetracker = "",
  px_to_mm.conversion = NULL,
  start_tracking.message = "default",
  start_tracking.match = "exact",
  subj.prefix = NULL,
  subj.suffix = NULL,
  timing.file = NULL,
  subset = "default",
  trial.exclude = c(),
  starttracking.message = NULL,
  starttracking.match = NULL
)
}
\arguments{
\item{file}{A file path to the raw data export file}

\item{eyetracker}{Which eye-tracker system was used to record data?}

\item{px_to_mm.conversion}{The conversion factor to go from
px pupil diameter to mm pupil diameter}

\item{start_tracking.message}{Message used in SMI experiment to
mark StartTracking inline}

\item{start_tracking.match}{Should the message string be an
"exact" match or a "pattern" match?}

\item{subj.prefix}{The unique pattern prefix (letter(s)
and/or symbol(s)) that comes before the subject number in the data file}

\item{subj.suffix}{The unique pattern suffix (letter(s) or
symbol(s)) that comes after the subject number in the data file}

\item{timing.file}{File location and name that contains timing
information for message markers}

\item{subset}{Which columns in the raw data export file do you want to keep}

\item{trial.exclude}{Specify if ther are any trials to exclude. Trial number}

\item{starttracking.message}{See start_tracking.message}

\item{starttracking.match}{See start_tracking.match}
}
\description{
This function converts a "messy" raw eye-tracker data file
to a "tidy" raw data file.
}
\details{
Different eye-trackers will save the raw data file using
completely different organizations and naming conventions
This function will create a "tidy" raw data file that has
a standard organization and naming convention,
regardless of which eye-tracker was used.
Supported Eye-Tracker Systems: Sensomotoric Instruments ("smi")
and SR-Research EyeLink ("eyelink")
}
