% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_upsample.R
\name{pupil_upsample}
\alias{pupil_upsample}
\title{Upsample pupil data}
\usage{
pupil_upsample(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will increase the sampling requency to 1000Hz
}
