% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eyelink.R
\name{read_eyelink}
\alias{read_eyelink}
\title{Convert SMI BeGaze for Analysis}
\usage{
read_eyelink(file = "", subset = "Time", trial.start = "TRIALID",
  eye = "both")
}
\arguments{
\item{file}{A file path to the raw data export file}

\item{subset}{Which columns in the raw data export file do you want to keep}

\item{trial.start}{Message used in SMI experiment to mark StartTracking inline}

\item{eye}{Do you want to inclue the "left", "right', or "both" eyes?}
}
\description{
This function takes a raw data export file from the SMI BeGaze software and converts it into a format that is easy to analyze
}
\examples{
read_eyelink(file = "path/filename.txt", subset = c(), track.start = "# Message: StartTracking.bmp", eye = "both", Hz = 1000)
}
\keyword{eyelink}
\keyword{read}
