% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_smi.R
\name{read_smi}
\alias{read_smi}
\title{Convert SMI BeGaze for Analysis}
\usage{
read_smi(file, trial.start = "# Message: StartTracking.bmp", eye = "both",
  subj.prefix = "-", trial.exclude = c())
}
\arguments{
\item{file}{A file path to the raw data export file}

\item{trial.start}{Message used in SMI experiment to mark StartTracking inline}

\item{eye}{Do you want to inclue the "left", "right', or "both" eyes?}

\item{subj.prefix}{The prefix that comes before the subject number in the data file (including "-")}

\item{trial.exclude}{Specify if ther are any trials to exclude. Trial number}
}
\description{
This function takes a raw data export file from the SMI BeGaze software and converts it into a format that is easy to analyze
}
\examples{
read_smi(file = "path/filename.txt", subset = c(), message.column = "columnName", track.start = "# Message: StartTracking.bmp", eye = "both")
}
\keyword{read}
\keyword{smi}
