% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_eye.R
\name{select_eye}
\alias{select_eye}
\title{Select eye to keep for analysis}
\usage{
select_eye(x, eye_use = "")
}
\arguments{
\item{x}{xframe}

\item{eye_use}{Which eye to use? Left or right}
}
\description{
If both eyes were recorded from, then this function will correlate the
timeseries pupil data from both eyes and select only one eye data to keep
for further preprocessing and output. This function will also remove
trials with too much missing data, specified by `missing_allowed`.
}
