% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.stimuli.R
\name{set.stimuli-deprecated}
\alias{set.stimuli-deprecated}
\title{Create a column of current stimulus period}
\usage{
set.stimuli(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will create a column that indexes the stimuli within each trial
}
\seealso{
\code{\link{pupillometry-deprecated}}
}
\keyword{internal}
