% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.stimuli.R
\name{set.stimuli}
\alias{set.stimuli}
\title{Indexing Function}
\usage{
set.stimuli(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will create a column that indexes the stimuli within each trial
}
\examples{
set.stimuli(x)
}
\keyword{set}
\keyword{stimuli}
