% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.timewindow.R
\name{set.timewindow}
\alias{set.timewindow}
\title{Indexing Function}
\usage{
set.timewindow(x, start = "", duration = 2000)
}
\arguments{
\item{x}{vector}

\item{start}{Message string that identifies the onset of a trial}

\item{duration}{Duration of time window in milliseconds}
}
\description{
This function uses specified start and/or end trial messages to correctly index the trial number. This makes it easier for pre-processing and later analyses
the 'dplyr' package is a dependency
}
\examples{
set.trial(x, start = "Fixation", duration = 2000)
}
\keyword{set}
\keyword{time}
\keyword{window}
