% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.timing.R
\name{set.timing-deprecated}
\alias{set.timing-deprecated}
\title{Set timing relative to trial-onset}
\usage{
set.timing(x, trialonset.message = "", ms.conversion = 1,
  pretrial.duration = 0, match = "exact")
}
\arguments{
\item{x}{dataframe}

\item{trialonset.message}{Message string that marks the start of a trial}

\item{ms.conversion}{Conversion factor to convert timing to milliseconds}

\item{pretrial.duration}{Duration of pre-trial baseline period in milliseconds}

\item{match}{Should the message string be an "exact" match or a "pattern" match?}
}
\description{
This sets the timing variable relative to the onset of each trial
}
\seealso{
\code{\link{pupillometry-deprecated}}
}
\keyword{internal}
