% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.trial.R
\name{set.trial-deprecated}
\alias{set.trial-deprecated}
\title{Set correct trial indexing}
\usage{
set.trial(x, startrecording.message = "", match = "exact")
}
\arguments{
\item{x}{dataframe}

\item{startrecording.message}{Message string that marks the start of a recorind}

\item{match}{Should the message string be an "exact" match or a "pattern" match?}
}
\description{
This sets the trials relative to a message string in the data
}
\seealso{
\code{\link{pupillometry-deprecated}}
}
\keyword{internal}
