% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_stimuli.R
\name{set_stimuli}
\alias{set_stimuli}
\title{Create a column of current stimulus period}
\usage{
set_stimuli(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will create a column that indexes the stimuli within each trial
}
\examples{
set_stimuli(x)
}
\keyword{set}
\keyword{stimuli}
