% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_timing.R
\name{set_timing}
\alias{set_timing}
\title{Set timing relative to some event onset}
\usage{
set_timing(
  x,
  onset_message = NULL,
  match = "exact",
  trial_onset.message = NULL,
  pretrial.duration = 0,
  trialonset.message = NULL,
  pre_trial.duration = NULL
)
}
\arguments{
\item{x}{dataframe}

\item{onset_message}{Message string that marks the start of the event}

\item{match}{Is message string an "exact" match or a "pattern" match?}

\item{trial_onset.message}{deprecated. Use onset_message.}

\item{pretrial.duration}{deprecated}

\item{trialonset.message}{deprecated}

\item{pre_trial.duration}{deprecated}
}
\description{
This sets the timing variable relative to the onset of an event
}
