% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_trial.R
\name{set_trial}
\alias{set_trial}
\title{Set correct trial indexing}
\usage{
set_trial(x, startrecording.message = "", match = "exact")
}
\arguments{
\item{x}{dataframe}

\item{startrecording.message}{Message string that marks the start of a recorind}

\item{match}{Should the message string be an "exact" match or a "pattern" match?}
}
\description{
This sets the trials relative to a message string in the data
}
\examples{
set_trial(x, start.trial = "Fixation", ms.conversion = 1000)
}
\keyword{set}
\keyword{timing}
