% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testResponse}
\alias{testResponse}
\title{testResponse}
\usage{
testResponse(response_data, testVar = "Response", forms = list())
}
\arguments{
\item{response_data}{A data.frame}

\item{testVar}{The variable to be tested (should be a column of `response_data`). Ignored if `forms` is given and complete.}

\item{forms}{An optional list with slots `full` and `reduced` containing the respective formulas.}
}
\description{
testResponse
}
