% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eyetracker.R
\name{tidy_eyetracker}
\alias{tidy_eyetracker}
\title{Messy to Tidy Raw Eye-Tracker Data File}
\usage{
tidy_eyetracker(file, eyetracker = "", trialmarker.message = "default",
  eye.recorded = "", subj.prefix = "default", subset = "default",
  trial.exclude = c())
}
\arguments{
\item{file}{A file path to the raw data export file}

\item{eyetracker}{Which eye-tracker system was used to record data?}

\item{trialmarker.message}{Message used in SMI experiment to mark StartTracking inline}

\item{eye.recorded}{Do you want to inclue the "left", "right', or "both" eyes?}

\item{subj.prefix}{The prefix that comes before the subject number in the data file (including "-")}

\item{subset}{Which columns in the raw data export file do you want to keep}

\item{trial.exclude}{Specify if ther are any trials to exclude. Trial number}
}
\description{
This function converts a "messy" raw eye-tracker data file to a "tidy" raw data file.
Different eye-trackers will save the raw data file using completely different organizations and naming conventions
This function will create a "tidy" raw data file that has a standard organization and naming convention, regardless of which eye-tracker was used.
Supported Eye-Tracker Systems: Sensomotoric Instruments ("smi") and SR-Research EyeLink ("eyelink")
}
\examples{
tidy_eyetracker(file = "path/filename.txt", subset = c(), message.column = "columnName", track.start = "# Message: StartTracking.bmp", eye = "both")
}
\keyword{tidy}
