% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.mean.R
\name{window.mean}
\alias{window.mean}
\title{Pupil Data Function}
\usage{
\method{window}{mean}(x, size = 0, Hz = 250)
}
\arguments{
\item{x}{vector of pupil data}

\item{size}{size of sliding time window in seconds (default: 0)}

\item{Hz}{the frequency at which pupil samples were collected (default: 250)}
}
\description{
This is a sliding time window function to calculate the mean of
each time window over a vector of pupil data. You can also use
this to calculate the overall mean by not specifying a window size
}
\examples{
window.mean(x, size = 30, Hz = 250)
}
\keyword{mean}
\keyword{window}
