% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.sd.R
\name{window.sd}
\alias{window.sd}
\title{Pupil Data Function}
\usage{
\method{window}{sd}(x, size = 0, Hz = 250)
}
\arguments{
\item{x}{vector of pupil data}

\item{size}{size of sliding time window in seconds (default: 0)}

\item{Hz}{the frequency at which pupil samples were collected (default: 250)}
}
\description{
This is a sliding time window function to calculate the standard
deviation of each time window over a vector of pupil data. You can
also use this to calculate the overall standard deviation by not.
specifying a window size.
}
\examples{
window.sd(x, size = 30, Hz = 250)
}
\keyword{sd}
\keyword{window}
