% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_downsample.R
\name{aoi_downsample}
\alias{aoi_downsample}
\title{aoi_downsample}
\usage{
aoi_downsample(x, bin.length = NULL, aoi.col = c())
}
\arguments{
\item{x}{dataframe}

\item{bin.length}{Length of bins to average}

\item{aoi.col}{Column names that contain aoi measures}
}
\description{
Downsample aoi data
}
\details{
This function will reduce the sampling fequency
}
