% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Pupil Preprocessing Function}
\usage{
downsample(x, Hz = "")
}
\arguments{
\item{x}{dataframe}

\item{Hz}{The sampling frequency to reduce pupil data to}
}
\description{
This function will reduce the sampling fequency to a specified Hz
}
\examples{
downsample(x, Hz = 100)
}
\keyword{downsample}
