% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes_detect.R
\name{eyes_detect}
\alias{eyes_detect}
\title{Detect Eyes}
\usage{
eyes_detect(x, include_bc = FALSE)
}
\arguments{
\item{x}{dataframe}

\item{include_bc}{logical, include the baseline corrected eyes}
}
\description{
Detect which eyes are in the data
https://dr-jt.github.io/pupillometry/ for more information.
}
\section{Output}{


This function returns the number of eyes recorded in the data
}

