% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.comparison.R
\name{plot.comparison}
\alias{plot.comparison}
\title{Testing Function}
\usage{
\method{plot}{comparison}(import, files = c(), labels = c(), trial,
  title = "")
}
\arguments{
\item{import}{Directory with preprocessed data files for comparison}

\item{files}{Vector of files to be compared}

\item{labels}{Vector of labels for plotting.}

\item{trial}{Which trial to filter on}
}
\description{
This function will print ggplot figures of na.removed, method.first = interpolate, and method.first = smooth
}
