% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pupil Data Function}
\usage{
preprocess(import = "", pattern = "*.txt", export = "",
  taskname = "", eyetracker = "", trialmarker.message = "default",
  trialonset.message = "", targetonset.message = "",
  eye.recorded = "", eye.use = "", eye.criteria = 0.9,
  pretrial.duration = "", bc.duration = "", velocity = "",
  margin = "", interpolate = FALSE, interpolate.type = "",
  interpolate.maxgap = Inf, smooth = FALSE, smooth.type = "",
  smooth.window = 5, downsample.Hz = "", bc = FALSE,
  subj.prefix = "default", subset = "default", trial.exclude = c())
}
\arguments{
\item{import}{Folder path to raw data files}

\item{pattern}{Pattern to look for in data files}

\item{export}{Folder path to export preprocessed data to}

\item{taskname}{Name of task - to be used in naming pre-processed files}

\item{eyetracker}{Which eye-tracker was used to record data}

\item{trialmarker.message}{Message used in SMI experiment to mark StartTracking inline}

\item{trialonset.message}{Message string that marks the start of a trial}

\item{targetonset.message}{Message string that marks the target onset for baseline correction}

\item{eye.recorded}{Do you want to inclue the "left", "right', or "both" eyes?}

\item{eye.use}{Which method of using left and right eye data? Average, missing, left, or right.}

\item{eye.criteria}{If using the average method, then what is the criteria of how highly the two eyes need to correlate?}

\item{pretrial.duration}{Duration of pre-trial baseline period in milliseconds}

\item{bc.duration}{PreTarget duration to use for baseline correction}

\item{velocity}{The velocity threshold for Blink detection}

\item{margin}{The margin before and after Blink onset and offset}

\item{interpolate}{Do you want to do a linear interpolation over missing values?}

\item{interpolate.type}{What type of interpolation to use? linear or cubic-spline}

\item{interpolate.maxgap}{Maximum number of NAs to interpolate over. Anything gaps over this value will not be interpolated.}

\item{smooth}{Do you want to apply a moving average smoothing function?}

\item{smooth.type}{The type of smoothing function to apply. hann or moving window average (mwa)}

\item{smooth.window}{Window size of smoothing function default is 5 milliseconds}

\item{downsample.Hz}{The frequency you want to downsample to}

\item{subj.prefix}{The prefix that comes before the subject number in the data file (including "-")}

\item{subset}{Which columns in the raw data export file do you want to keep}

\item{trial.exclude}{Specify if ther are any trials to exclude. Trial number}
}
\description{
This function will perform preprocessing on an entire folder of data files. It takes as input a specified folder path
and outputs the preprocessed data to a specified folder path
}
\examples{


}
\keyword{preprocess}
