% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess pupil data using this single function}
\usage{
preprocess(import = NULL, pattern = "*.txt", taskname = NULL,
  subj.prefix = NULL, subj.suffix = NULL, output = NULL,
  gazedata.include = FALSE, eyetracker = NULL, hz = NULL,
  eye.recorded = NULL, eye.use = NULL,
  startrecording.message = "default", startrecording.match = "exact",
  trialonset.message = NULL, trialonset.match = "exact",
  pretrial.duration = NULL, missing.allowed = 1, interpolate = NULL,
  interpolate.maxgap = Inf, smooth = NULL, smooth.window = 5,
  method.first = NULL, bc = NULL, bc.duration = NULL,
  baselineoffset.message = NULL, baselineoffset.match = "exact",
  downsample.binlength = NULL, subset = "default",
  trial.exclude = c())
}
\arguments{
\item{import}{Folder path to raw data files}

\item{pattern}{Pattern to look for in data files}

\item{taskname}{Name of task - to be used in naming pre-processed files}

\item{subj.prefix}{The unique pattern prefix (letter(s) and/or symbol(s)) that comes before the subject number in the data file}

\item{subj.suffix}{The unique pattern suffix (letter(s) or symbol(s)) that comes after the subject number in the data file}

\item{output}{Folder path to output preprocessed data to}

\item{gazedata.include}{Logical. Include columns for x and y coordinates of eye gaze? (Default: FALSE)}

\item{eyetracker}{Which eye-tracker was used to record data}

\item{hz}{At which frequency was pupil data sampled at? (only required for interpolation and smoothing)}

\item{eye.recorded}{Do you want to inclue the "left", "right', or "both" eyes?}

\item{eye.use}{Which eye to use? Left or right}

\item{startrecording.message}{Message used in SMI experiment to mark StartTracking inline}

\item{startrecording.match}{Should the message string be an "exact" match or a "pattern" match?}

\item{trialonset.message}{Message string that marks the start of a trial}

\item{trialonset.match}{Should the message string be an "exact" match or a "pattern" match?}

\item{pretrial.duration}{Duration of pre-trial baseline period in milliseconds}

\item{missing.allowed}{What proportion of missing data is allowed, on a trial-by-trial basis? (Default: 1)}

\item{interpolate}{What type of interpolation to use? linear or cubic-spline}

\item{interpolate.maxgap}{Maximum number of NAs to interpolate over. Anything gaps over this value will not be interpolated.}

\item{smooth}{The type of smoothing function to apply. hann or moving window average (mwa)}

\item{smooth.window}{Window size of smoothing function default is 5 milliseconds}

\item{method.first}{Should "smooth" or "interpolate" be applied first? (default: NULL)}

\item{bc}{Do you want to use "subtractive" or "divisive" baseline correction? (default: "subtractive")}

\item{bc.duration}{Duration baseline period(s) to use for correction}

\item{baselineoffset.message}{Message string(s) that marks the offset of baseline period(s)}

\item{baselineoffset.match}{Message string(s) that marks the offset of baseline period(s)}

\item{downsample.binlength}{Length of bins to average (default: NULL)}

\item{subset}{Which columns in the raw data output file do you want to keep}

\item{trial.exclude}{Specify if ther are any trials to exclude. Trial number}
}
\description{
This function will perform preprocessing on an entire folder of data files. It takes as input a specified folder path
and outputs the preprocessed data to a specified folder path
}
\examples{


}
\keyword{preprocess}
