% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pupil Data Function}
\usage{
preprocess(import = "", pattern = "*.txt", output = NULL,
  export = "", taskname = "", eyetracker = "", eye.recorded = "",
  eye.use = "", hz = "", startrecording.message = "default",
  startrecording.match = "exact", trialonset.message = "",
  pretrial.duration = "", velocity = "", margin = "",
  interpolate = FALSE, interpolate.type = "",
  interpolate.maxgap = Inf, smooth = FALSE, smooth.type = "",
<<<<<<< HEAD
  smooth.window = 5, bc = FALSE, baselineoffset.message = "",
  bc.duration = "", downsample.binlength = "", subj.prefix = NULL,
  subj.suffix = NULL, subset = "default", trial.exclude = c())
=======
  smooth.window = 5, method.first = NULL, bc = FALSE,
  baselineoffset.message = "", bc.duration = "",
  downsample.binlength = "", subj.prefix = "default",
  subset = "default", trial.exclude = c())
>>>>>>> reorder
}
\arguments{
\item{import}{Folder path to raw data files}

\item{pattern}{Pattern to look for in data files}

\item{output}{Folder path to output preprocessed data to}

\item{export}{Folder path to export preprocessed data to}

\item{taskname}{Name of task - to be used in naming pre-processed files}

\item{eyetracker}{Which eye-tracker was used to record data}

\item{eye.recorded}{Do you want to inclue the "left", "right', or "both" eyes?}

\item{eye.use}{Which eye to use? Left or right}

\item{hz}{At which frequency was pupil data sampled at? (only required for interpolation and smoothing)}

\item{startrecording.message}{Message used in SMI experiment to mark StartTracking inline}

\item{startrecording.match}{Should the message string be an "exact" match or a "pattern" match?}

\item{trialonset.message}{Message string that marks the start of a trial}

\item{pretrial.duration}{Duration of pre-trial baseline period in milliseconds}

\item{velocity}{The velocity threshold for Blink detection}

\item{margin}{The margin before and after Blink onset and offset}

\item{interpolate}{Do you want to do a linear interpolation over missing values?}

\item{interpolate.type}{What type of interpolation to use? linear or cubic-spline}

\item{interpolate.maxgap}{Maximum number of NAs to interpolate over. Anything gaps over this value will not be interpolated.}

\item{smooth}{Do you want to apply a moving average smoothing function?}

\item{smooth.type}{The type of smoothing function to apply. hann or moving window average (mwa)}

\item{smooth.window}{Window size of smoothing function default is 5 milliseconds}

\item{method.first}{Should "smooth" or "interpolate" be applied first? (default: NULL)}

\item{bc}{Logical. Do baseline correction?}

\item{baselineoffset.message}{Message string(s) that marks the offset of baseline period(s)}

\item{bc.duration}{Duration baseline period(s) to use for correction}

\item{downsample.binlength}{Length of bins to average}

\item{subj.prefix}{The unique pattern prefix (letter(s) and/or symbol(s)) that comes before the subject number in the data file}

\item{subj.suffix}{The unique pattern suffix (letter(s) or symbol(s)) that comes after the subject number in the data file}

\item{subset}{Which columns in the raw data output file do you want to keep}

\item{trial.exclude}{Specify if ther are any trials to exclude. Trial number}
}
\description{
This function will perform preprocessing on an entire folder of data files. It takes as input a specified folder path
and outputs the preprocessed data to a specified folder path
}
\examples{


}
\keyword{preprocess}
