% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.baselinecorrect.R
\name{pupil.baselinecorrect}
\alias{pupil.baselinecorrect}
\title{Pupil Data Function}
\usage{
pupil.baselinecorrect(x, baselineoffset.message = "",
  bc.duration = 200)
}
\arguments{
\item{x}{dataframe}

\item{baselineoffset.message}{Message string(s) that marks the offset of baseline period(s)}

\item{bc.duration}{Duration baseline period(s) to use for correction}
}
\description{
This function applies a pre-trial baseline correction on the data
}
\examples{
pupil.baselinecorrect(file = "path/filename", baseline.duration = 2000, start.trial = "# Message: Target")
}
\keyword{baseline}
