% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.cor.R
\name{pupil.cor}
\alias{pupil.cor}
\title{Preprocessnig Function}
\usage{
pupil.cor(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function is for preprocessing baseline pupil data. It takes as input a data file that has been processed
with the .convert() functions. The output of this function is a preprocessed data file. Preprocessing steps can
include interpolation and smoothing
}
\examples{
pupil.cor(x)
}
\keyword{eye}
