% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.interpolate.R
\name{pupil.interpolate}
\alias{pupil.interpolate}
\title{Pupil Preprocessing Function}
\usage{
pupil.interpolate(x, type = "cubic-spline", maxgap = Inf, hz = "")
}
\arguments{
\item{x}{dataframe}

\item{type}{What type of interpolation to use? (default: cubic-spline)}

\item{maxgap}{Maximum number of NAs to interpolate over. Anything gaps over this value will not be interpolated.}

\item{hz}{The recording frequency (used to calculate window size)}
}
\description{
This function performs a linear interpolation over missing values.
The 'zoo' package is a dependency
}
\examples{
pupil.interpolate(x, type = "cubic-spline", eye.recorded = "both")
}
\keyword{interpolate}
