% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.interpolate.R
\name{pupil.interpolate}
\alias{pupil.interpolate}
\title{Pupil Preprocessing Function}
\usage{
pupil.interpolate(x, type = "cubic-spline", eye.recorded = "")
}
\arguments{
\item{x}{xframe}

\item{type}{What type of interpolation to use? (default: cubic-spline)}

\item{eye.recorded}{Is there pupil x for the left, right, or both eyes?}
}
\description{
This function performs a linear interpolation over missing values.
The 'zoo' package is a dependency
}
\examples{
pupil.interpolate(x, type = "cubic-spline", eye.recorded = "both")
}
\keyword{interpolate}
