% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.smooth.R
\name{pupil.smooth}
\alias{pupil.smooth}
\title{Pupil Preprocessing Function}
\usage{
pupil.smooth(x, type = "hann", window = 11, hz = "", eye.recorded = "")
}
\arguments{
\item{x}{dataframe}

\item{type}{The type of smoothing function to apply. Hann or moving window average}

\item{window}{The size of the smoothing window (default = 11). Value is in milliseconds}

\item{hz}{The recording frequency (used to calculate window size)}

\item{eye.recorded}{Is there pupil data for the left, right, or both eyes?}
}
\description{
This function performs a smoothing function over an array of data
The 'zoo' package is a dependency
}
\examples{
pupil.smooth(x, type = "hann", window = 11, eye.recorded = "both")
}
\keyword{smooth}
