% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_baselinecorrect.R
\name{pupil_baselinecorrect}
\alias{pupil_baselinecorrect}
\title{Apply baseline correction}
\usage{
pupil_baselinecorrect(x, baselineoffset.message = "",
  bc.duration = 200, bc.type = "subtractive", match = "exact")
}
\arguments{
\item{x}{dataframe}

\item{baselineoffset.message}{Message string(s) that marks the offset of baseline period(s)}

\item{bc.duration}{Duration baseline period(s) to use for correction}

\item{bc.type}{Do you want to use "subtractive" or "divisive" baseline correction? (default: "subtractive")}

\item{match}{Should the message string be an "exact" match or a "pattern" match?}
}
\description{
This function applies a pre-trial baseline correction on the data
}
\examples{
pupil_baselinecorrect(file = "path/filename", baseline.duration = 2000, start.trial = "# Message: Target")
}
\keyword{baseline}
