% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_baselinecorrect.R
\name{pupil_baselinecorrect}
\alias{pupil_baselinecorrect}
\title{Apply baseline correction}
\usage{
pupil_baselinecorrect(
  x,
  bc_onset.message = "",
  pre.duration = 200,
  type = "subtractive",
  match = "exact"
)
}
\arguments{
\item{x}{dataframe}

\item{bc_onset.message}{Message string(s) that marks the offset of
baseline period(s)}

\item{pre.duration}{Duration baseline period(s) to use for correction}

\item{type}{Do you want to use "subtractive" or "divisive"
baseline correction? (default: "subtractive")}

\item{match}{Is the message string an "exact" match or a "pattern" match?}
}
\description{
This function applies a baseline correction on the data
}
