% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_bin.R
\name{pupil_bin}
\alias{pupil_bin}
\title{Bin pupil data}
\usage{
pupil_bin(x, bin_length = NULL)
}
\arguments{
\item{x}{dataframe.}

\item{bin_length}{Length of bins in milliseconds.}
}
\description{
Bin pupil data into smaller time-bins. This is useful to perform time-series
analyses or to just reduce the size of the data. See
https://dr-jt.github.io/pupillometry/ for more information.
}
\section{Output}{


Reduces the number of rows into discrete time-bins. Changes the values in
Time and pupil columns.
}

