% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_cor.R
\name{pupil_cor}
\alias{pupil_cor}
\title{Add column that is the correlation between left and rigth eyes}
\usage{
pupil_cor(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function is for preprocessing baseline pupil data. It takes as input a data file that has been processed
with the .convert() functions. The output of this function is a preprocessed data file. Preprocessing steps can
include interpolation and smoothing
}
\keyword{eye}
