% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_cor.R
\name{pupil_cor}
\alias{pupil_cor}
\title{Add column that is the correlation between left and rigth eyes}
\usage{
pupil_cor(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will simply calculate the correlation between
left and right eyes
}
