% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_deblink.R
\name{pupil_deblink}
\alias{pupil_deblink}
\title{Deblink}
\usage{
pupil_deblink(x, extend = 0, plot = FALSE)
}
\arguments{
\item{x}{dataframe.}

\item{extend}{How many milliseconds to extend blinks
before and after blink detection.}

\item{plot}{Logical. Inspect a plot of how pupil values changed?}
}
\description{
Deblink pupil data with the option to extend blinks before and after
blink detection.
}
\details{
See https://dr-jt.github.io/pupillometry/index.html for more information.
}
\section{Output}{


Changes values in column containing pupil data.
}

\section{Plot inspection}{


To inspect how the preprocesing step changed pupil size values,
use `plot = TRUE`.

Warning: this will create a separate plot for every trial and therefore can
be time consuming and overwhelming. The plot argument is meant for initial
exploratory steps to determine the appropriate preprocessing parameters.
}

