% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_deblink.R
\name{pupil_deblink}
\alias{pupil_deblink}
\title{Evaluate the amount of missing data}
\usage{
pupil_deblink(x, extend = 100)
}
\arguments{
\item{x}{dataframe}

\item{extend}{How many milliseconds to extend blinks
before and after blink detection}
}
\description{
This function will deblink pupil data with the option to extend
    blinks before and after blink detection
}
\keyword{deblink}
