% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_downsample.R
\name{pupil_downsample}
\alias{pupil_downsample}
\title{Downsample pupil data}
\usage{
pupil_downsample(x, bin.length = NULL, bc = FALSE)
}
\arguments{
\item{x}{dataframe}

\item{bin.length}{Length of bins to average}

\item{bc}{Logical. Was baseline correction done on the data? Is there a column labeled Pupil_Diameter_bc.mm?}
}
\description{
This function will reduce the sampling fequency
}
\examples{
pupil_downsample(x, bin.length = 100)
}
\keyword{downsample}
