% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_downsample.R
\name{pupil_downsample}
\alias{pupil_downsample}
\title{Downsample pupil data}
\usage{
pupil_downsample(x, bin.length = NULL, id = "Subject")
}
\arguments{
\item{x}{dataframe}

\item{bin.length}{Length of bins to average}

\item{id}{Column name that identifies subject number}
}
\description{
This function will reduce the sampling fequency
}
