% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_missing.R
\name{pupil_missing}
\alias{pupil_missing}
\title{Remove trials with too much missing data}
\usage{
pupil_missing(x, missing_allowed = 1, velocity = "", margin = "")
}
\arguments{
\item{x}{dataframe}

\item{missing_allowed}{What proportion of missing data is allowed, per trial?}
}
\description{
This function will remove trials that have too much missing data
}
