% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_outliers.R
\name{pupil_outliers}
\alias{pupil_outliers}
\title{Outlier detection and removal for pupil data}
\usage{
pupil_outliers(
  x,
  threshold = 3,
  remove = TRUE,
  plot = FALSE,
  plot_trial = "all",
  histogram = FALSE,
  binwidth = 0.1
)
}
\arguments{
\item{x}{dataframe.}

\item{threshold}{Number of standard deviations from the mean to use as the outlier threshold. Default is 3.}

\item{remove}{Logical. Whether to remove outliers from the data. Default is TRUE.}

\item{plot}{Logical. Whether to create a plot showing the distribution of pupil size values with outlier thresholds. Default is FALSE.}

\item{plot_trial}{What trial(s) to plot if `plot = TRUE`. Default is "all".}

\item{histogram}{Logical. Inspect a histogram of how pupil values changed? Default is FALSE.}

\item{binwidth}{Numeric. Bin width for the histogram if `histogram = TRUE`. Default is 0.1.}
}
\description{
This function detects and removes outliers in pupil size data using
standard deviation from the mean. It is designed to identify and handle
extreme values that may be due to measurement errors or other anomalies
in the data.
}
\section{Output}{


Changes values in the column that contains pupil data to NA if they are
identified as outliers based on the specified threshold.
}

\section{Standard deviation method}{


The function calculates the mean and standard deviation of the pupil size
data across all observations for each participant. It then identifies
outliers as values that are more than a specified number of standard
deviations away from the mean. The default threshold is set to 3, meaning
that any value more than 3 standard deviations from the mean will be
considered an outlier.
}

\section{Plot inspection}{


To inspect how the preprocessing step changed pupil size values,
use `plot = TRUE`. This will create a plot showing the distribution of pupil
size values with lines indicating the mean and the outlier thresholds. This
can help you visually assess the impact of the outlier removal process.
}

