% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_plot.R
\name{pupil_plot}
\alias{pupil_plot}
\title{Plot and inspect pupil data}
\usage{
pupil_plot(x, y, trial = "all", sub_title = "")
}
\arguments{
\item{x}{data before preprocessing step.}

\item{y}{data after preprocessing step.}

\item{trial}{what trial(s) to plot. default = "all"}

\item{sub_title}{subtitle for plot. default = ""}
}
\description{
Plot and inspect how preprocessing steps are changing the pupil data.
An exploratory function to help determine preprocessing parameters. See
https://dr-jt.github.io/pupillometry/ for more information.
}
\section{Output}{


Plots.
}

\section{Usage}{


This function is primarily used within other preprocessing functions.
(e.g., `pupil_smooth(plot = TRUE)`).
}

