% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_smooth.R
\name{pupil_smooth}
\alias{pupil_smooth}
\title{Apply smoothing/low frequency filter}
\usage{
pupil_smooth(x, type = "hann", n = NULL, window = NULL, hz = NULL)
}
\arguments{
\item{x}{dataframe}

\item{type}{The type of smoothing function to apply. "hann" or "mwa"}

\item{n}{Number of samples to smooth over
(length of hanning filter for "hann" and window size for "mwa")}

\item{window}{deprecated. See n}

\item{hz}{deprecated. See n}
}
\description{
This function performs a smoothing function over an array of x
The 'zoo' package is a dependency
}
