% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.downsample.R, R/plot.comparison.R,
%   R/pupil.baselinecorrect.R, R/pupil.cor.R, R/pupil.downsample.R,
%   R/pupil.eye.R, R/pupil.interpolate.R, R/pupil.missing.R, R/pupil.smooth.R,
%   R/pupillometry-deprecated.r, R/set.stimuli.R, R/set.timing.R, R/set.trial.R
\name{pupillometry-deprecated}
\alias{pupillometry-deprecated}
\alias{aoi.downsample}
\alias{plot.comparison}
\alias{pupil.baselinecorrect}
\alias{pupil.cor}
\alias{pupil.downsample}
\alias{pupil.eye}
\alias{pupil.interpolate}
\alias{pupil.missing}
\alias{pupil.smooth}
\alias{set.stimuli}
\alias{set.timing}
\alias{set.trial}
\title{Deprecated functions in package \pkg{pupillometry}.}
\usage{
aoi.downsample(x, bin.length = NULL, aoi_col = c())

\method{plot}{comparison}(import, files = c(), labels = c(), trial,
  title = "", legend.title = "Labels")

pupil.baselinecorrect(x, baselineoffset.message = "",
  bc.duration = 200, bc.type = "subtractive", match = "exact")

pupil.cor(x)

pupil.downsample(x, bin.length = NULL, bc = FALSE)

pupil.eye(x, eye.recorded = "", eye.use = "",
  gazedata.include = FALSE)

pupil.interpolate(x, type = "cubic-spline", maxgap = Inf, hz = "")

pupil.missing(x, missing.allowed = 1, velocity = "", margin = "")

pupil.smooth(x, type = "hann", window = 5, hz = "")

set.stimuli(x)

set.timing(x, trialonset.message = "", ms.conversion = 1,
  pretrial.duration = 0, match = "exact")

set.trial(x, startrecording.message = "", match = "exact")
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\section{\code{aoi.downsample}}{

For \code{aoi.downsample}, use \code{\link{aoi_downsample}}.
}

\section{\code{plot.comparison}}{

For \code{plot.comparison}, use \code{\link{plot_comparison}}.
}

\section{\code{pupil.baselinecorrect}}{

For \code{pupil.baselinecorrect}, use \code{\link{pupil_baselinecorrect}}.
}

\section{\code{pupil.cor}}{

For \code{pupil.cor}, use \code{\link{pupil_cor}}.
}

\section{\code{pupil.downsample}}{

For \code{pupil.downsample}, use \code{\link{pupil_downsample}}.
}

\section{\code{pupil.eye}}{

For \code{pupil.eye}, use \code{\link{pupil_eye}}.
}

\section{\code{pupil.interpolate}}{

For \code{pupil.interpolate}, use \code{\link{pupil_interpolate}}.
}

\section{\code{pupil.missing}}{

For \code{pupil.missing}, use \code{\link{pupil_missing}}.
}

\section{\code{pupil.smooth}}{

For \code{pupil.smooth}, use \code{\link{pupil_smooth}}.
}

\section{\code{set.stimuli}}{

For \code{set.stimuli}, use \code{\link{set_stimuli}}.
}

\section{\code{set.timing}}{

For \code{set.timing}, use \code{\link{set_timing}}.
}

\section{\code{set.trial}}{

For \code{set.trial}, use \code{\link{set_trial}}.
}

\examples{
aoi.downsample(x, bin.length = 100)

pupil.baselinecorrect(file = "path/filename", baseline.duration = 2000, start.trial = "# Message: Target")
pupil.cor(x)
pupil.downsample(x, bin.length = 100)

pupil.interpolate(x, type = "cubic-spline", eye.recorded = "both")
pupil.missing(x)
pupil.smooth(x, type = "hann", window = 11, eye.recorded = "both")
set.stimuli(x)
set.timing(x, start.trial = "Fixation", ms.conversion = 1000)
set.trial(x, start.trial = "Fixation", ms.conversion = 1000)
}
\keyword{baseline}
\keyword{downsample}
\keyword{eye}
\keyword{internal}
\keyword{interpolate}
\keyword{plot}
\keyword{preprocess}
\keyword{pupil}
\keyword{set}
\keyword{smooth}
\keyword{stimuli}
\keyword{timing}
