% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.timing.R
\name{set.timing}
\alias{set.timing}
\title{Pupil Data Function}
\usage{
set.timing(x, start.trial = "", ms.conversion = 1,
  pretrial.duration = 0)
}
\arguments{
\item{x}{dataframe}

\item{start.trial}{Message string that marks the start of a trial}

\item{ms.conversion}{Conversion factor to convert timing to milliseconds}

\item{pretrial.duration}{Duration of pre-trial baseline period in milliseconds}
}
\description{
This sets the timing variable relative to the onset of each trial
}
\examples{
set.timing(x, start.trial = "Fixation", ms.conversion = 1000)
}
\keyword{set}
\keyword{timing}
