% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.trial.R
\name{set.trial}
\alias{set.trial}
\title{Pupil Data Function}
\usage{
set.trial(x, startrecording.message = "")
}
\arguments{
\item{x}{dataframe}

\item{startrecording.message}{Message string that marks the start of a recorind}
}
\description{
This sets the trials relative to a message string in the data
}
\examples{
set.trial(x, start.trial = "Fixation", ms.conversion = 1000)
}
\keyword{set}
\keyword{timing}
